+++ internal/attestation/types/types.go
@@ -1,0 +1,7 @@
+package attestationtypes
+
+// ID holds the identifiers of a node.
+type ID struct {
+	Cluster []byte `json:"cluster"`
+	Owner   []byte `json:"owner"`
+}